+++ example/lib/app/app_module.dart
@@ -1,7 +1,14 @@
+<<<<<<< HEAD
 import 'package:bloc_pattern/bloc_pattern.dart';
 import 'package:flutter/material.dart';
 import 'package:example/app/app_widget.dart';
 import 'package:example/app/app_bloc.dart';
+=======
+import 'package:example/app/app_bloc.dart';
+import 'package:bloc_pattern/bloc_pattern.dart';
+import 'package:flutter/material.dart';
+import 'package:example/app/app_widget.dart';
+>>>>>>> 2f431f953aed90d425f32232a3317fd7f82f3bb4
 
 class AppModule extends ModuleWidget {
   @override