+++ lib/slidy.dart
@@ -1,7 +1,6 @@
 export 'src/modules/modules.dart';
 
 export 'package:slidy/src/command/command_base.dart';
-export 'package:slidy/src/command/generate_command.dart';
 export 'package:slidy/src/command/install_command.dart';
 export 'package:slidy/src/command/start_command.dart';
 export 'package:slidy/src/command/uninstall_command.dart';